[ring] Use gulong instead of guint for row/col indices
Bug 599444 - Scrollback index type mess

+++ src/buffer.h
@@ -36,7 +36,7 @@ typedef GByteArray VteBuffer;
 #define _vte_buffer_length(B)			((B)->len)
 #define _vte_buffer_consume(B, length)		g_byte_array_remove_range (B, 0, length)
 #define _vte_buffer_clear(B)			g_byte_array_set_size (B, 0)
-#define _vte_buffer_set_minimum_size(B, length)	g_byte_array_set_size (B, MAX ((length), (B)->len))
+#define _vte_buffer_set_minimum_size(B, length)	g_byte_array_set_size (B, (guint) MAX ((gint) (length), (gint) (B)->len))
 
 G_END_DECLS
